Open source the ETA C API RWF Encoder/Decoder and Transport .  The ESDK1.2.0.L1 release contains the following:
* Release of the ETA C TCP-based transports and Sequenced Multicast transport
* Release of the ETA C RWF Encoder and Decoder
* CMake build scripts for both ETA C and EMA C++ APIs.  For more information about the usage of these scripts, please look at the READMEs.

+++ Cpp-C/Ema/Src/Access/Impl/FilterListEncoder.cpp
@@ -235,7 +235,7 @@ void FilterListEncoder::complete()
 	_containerComplete = true;
 }
 
-void FilterListEncoder::totalCountHint( UInt32 totalCountHint )
+void FilterListEncoder::totalCountHint( UInt8 totalCountHint )
 {
 	if ( !_containerInitialized )
 	{
